requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
pinecone-client>=4.0.0
numpy>=1.21.0,<1.25.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
openpyxl>=3.1.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
datasets>=2.14.0
accelerate>=0.20.0
evaluate>=0.4.0
